+++ auth-vue/src/index.ts
@@ -1,5 +1,5 @@
 export * from "./requests/auth.ts";
-export * from "./types.ts";
+export * from "./requests/types.ts";
 export { default as LoginPage } from "./components/LoginPage.vue";
 export { default as RegisterPage } from "./components/RegisterPage.vue";
 export { default as ForgotPasswordPage } from "./components/ForgotPasswordPage.vue";
